% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_struct_example}
\alias{create_struct_example}
\title{create_struct_example}
\usage{
create_struct_example()
}
\value{
A external pointer object is returned containing a pointer to the Arrow array with
a pointer to the schema tucked away in the \sQuote{tag} component as is common for
\pkg{nanoarrow} objects
}
\description{
Create a Simple \sQuote{struct} Arrow Object
}
\details{
A two-column \sQuote{struct} is created and returned, containing respectively an
integer and a floating point column each with one missing value
}
\examples{
s <- create_struct_example()
s                 # pretty-printed in full details thanks to 'nanoarrow' S3 methods
as.data.frame(s)  # converted thanks to 'nanoarrow' S3 methods
}
