% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{show_array}
\alias{show_array}
\alias{show_schema}
\alias{array_xptr_set_schema}
\alias{array_xptr_get_schema}
\title{Display Arrow Structures Array or Schema}
\usage{
show_array(sxparr)

show_schema(sxparr)

array_xptr_set_schema(array_xptr, schema_xptr)

array_xptr_get_schema(array_xptr)
}
\arguments{
\item{sxparr}{An external pointer to an \sQuote{Arrow Array} also containing
a pointer to the corresponding \sQuote{Arrow Schema} in its \sQuote{tag} slot}

\item{array_xptr}{A external pointer to an Arrow array data structure}

\item{schema_xptr}{A external pointer to an Arrow schema data structure}
}
\value{
Nothing is returned, the functions are invoked for the side effect
of printing
}
\description{
Helper functions showing elements of \sQuote{Arrow Array} or
\sQuote{Arrow Schema} objects provided via an external pointer
}
